remove uneeded dependencies, ensure immutability

+++ src/components/Idea/Idea.tsx
@@ -1,25 +1,35 @@
-import { useState } from 'react';
-import { format } from "date-fns";
-import type { IdeaType } from '../../types/index';
+import { useState, useRef, useEffect } from 'react';
+import { format } from 'date-fns';
+import type { IdeaType, IdeaText } from '../../types/index';
 
 type IdeaProps = {
   idea: IdeaType;
-  handleEdit: (id: number, ideaObj: IdeaType['content']) => void;
+  handleEdit: (id: number, ideaObj: IdeaText) => void;
   handleDelete?: (id: number) => void;
-  autoFocus: true;
-  handleCreate: (ideaObj: IdeaType['content']) => void;
+  autoFocus: boolean;
+  handleCreate: (ideaObj: IdeaText) => void;
 };
 
 export const Idea = ({ handleCreate, handleEdit, idea, handleDelete, autoFocus }: IdeaProps) => {
   const { createdAt, modifiedAt } = idea.content;
   const [title, setTitle] = useState<string>(idea.content.title);
   const [description, setDescription] = useState<string>(idea.content.description);
-  console.log('idea ==>', idea);
+
+  const inputRef = useRef<HTMLInputElement>(null);
+
+  /* Keep the input place holder in focus when the idea is created or edited. */
+  // biome-ignore lint/correctness/useExhaustiveDependencies: <explanation>
+  useEffect(() => {
+    if(autoFocus && inputRef.current){
+      inputRef.current.focus();
+    }
+  }, [autoFocus, idea])
+
   const onSubmit = () => {
     const ideaObj = {
       title,
       description,
-      createdAt: new Date(),
+      createdAt: new Date()
     };
     handleCreate(ideaObj);
 
@@ -28,19 +38,18 @@ export const Idea = ({ handleCreate, handleEdit, idea, handleDelete, autoFocus }
   };
 
   const onEdit = () => {
-    const ideaObj = {
+    const contentObj = {
       title,
       description,
-      modifiedAt: new Date(),
     };
 
-    console.log('On Edit', ideaObj);
-
-    handleEdit(idea.id, ideaObj);
+    handleEdit(idea.id, contentObj);
   };
 
   const onDelete = () => {
-    handleDelete(idea.id);
+    if(handleDelete) {
+      handleDelete(idea.id);
+    }
   };
 
   return (
@@ -52,22 +61,22 @@ export const Idea = ({ handleCreate, handleEdit, idea, handleDelete, autoFocus }
         data-testid="idea-item"
         name={`idea${idea.id}title`}
         onChange={(e) => setTitle(e.target.value)}
-        autoFocus={autoFocus}
+        ref={inputRef}
         required
       />
       <textarea
         value={description}
         onChange={(e) => setDescription(e.target.value)}
-        rows="2"
-        cols="50"
+        rows={2}
+        cols={50}
         aria-label={`idea${idea.id}description`}
         maxLength={140}
         required
       />
       {modifiedAt ? <p>last modified at {format(modifiedAt, 'dd-MM-yyyy HH:mm:ss')}</p> : ''}
       {createdAt ? <p>Last created at {format(createdAt, 'dd-MM-yyyy HH:mm:ss')}</p> : ''}
-      {createdAt ? <button onClick={onEdit}>Edit</button> : <button onClick={onSubmit}>Add</button>}
-      {handleDelete ? <button onClick={onDelete}>delete</button> : ''}
+      {createdAt ? <button type="button" onClick={onEdit}>Edit</button> : <button type="button" onClick={onSubmit}>Add</button>}
+      {handleDelete ? <button type="button" onClick={onDelete}>delete</button> : ''}
     </div>
   );
 };
